requirements.txt
replicate
flask
gunicorn
google-generativeai
